# documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "description": "",
        "title": "Road Protect API",
        "version": "1.4.1"
    },
    "paths": {
...
        "/api/v1/document/{documentId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Document"
                        }
                    },
                    "400": {
                        "description": "E044: Could not find Document with id: documentId"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Documents & Files"
                ],
                "operationId": "DocumentController_getDocument",
                "summary": "Get document by DocumentId"
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateDocumentDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Document"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Documents & Files"
                ],
                "operationId": "DocumentController_updateDocument",
                "summary": "Update document by DocumentId"
            }
        },
...
        "/api/v1/infringement/{infringementId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "infringementId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Infringement"
                        }
                    },
                    "400": {
                        "description": "E037: Could not find Infringement with id: infringementId"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Infringements"
                ],
                "operationId": "InfringementController_getInfringement",
                "summary": "Get infringement by Infringement Id"
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "infringementId",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateInfringementDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Infringement"
                        }
                    },
                    "400": {
                        "description": " * Invalid infringement status update. Cannot go from current to next.\n * E037: Could not find Infringement with id: infringementId"
                    },
                    "500": {
                        "description": "Database Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Infringements"
                ],
                "operationId": "InfringementController_updateInfringement",
                "summary": "Update infringement"
            }
        },
...
    },
    "definitions": {
...
        "Document": {
            "title": "Document",
            "properties": {
                "account": {
                    "description": "Account",
                    "type": "object"
                },
                "contract": {
                    "description": "Contract",
                    "type": "object"
                },
                "createdAt": {
                    "type": "string"
                },
                "documentId": {
                    "type": "number"
                },
                "fileDirectory": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "generatedDocument": {
                    "description": "GeneratedDocument",
                    "type": "object"
                },
                "infringement": {
                    "description": "Infringement",
                    "type": "object"
                },
                "leaseContract": {
                    "description": "LeaseContract",
                    "type": "object"
                },
                "manualPayment": {
                    "description": "ManualPayment",
                    "type": "object"
                },
                "nominationMerged": {
                    "description": "Nomination[]",
                    "type": "object"
                },
                "nominationRedirection": {
                    "description": "Nomination[]",
                    "type": "object"
                },
                "ocr": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/OcrDetails"
                        },
                        {
                            "description": "Results from contract OCR"
                        }
                    ]
                },
                "relations": {
                    "description": "AccountRelation",
                    "type": "object"
                },
                "storageName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicleDocuments": {
                    "description": "Vehicle Documents [Registration] (not used for much right now)",
                    "type": "object"
                }
            },
            "required": [
                "createdAt",
                "updatedAt",
                "documentId",
                "storageName",
                "fileName",
                "fileDirectory",
                "ocr",
                "contract",
                "leaseContract",
                "vehicleDocuments",
                "infringement",
                "manualPayment",
                "nominationMerged",
                "nominationRedirection",
                "account",
                "generatedDocument",
                "relations"
            ],
            "type": "object"
        },
...
        "Infringement": {
            "title": "Infringement",
            "properties": {
                "amountDue": {
                    "type": "string"
                },
                "approvedDate": {
                    "type": "string"
                },
                "brn": {
                    "type": "string"
                },
                "caseNumber": {
                    "type": "string"
                },
                "contract": {
                    "description": "Contract",
                    "type": "object"
                },
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "description": "Document",
                    "type": "object"
                },
                "externalChangeDate": {
                    "type": "string"
                },
                "infringementApproval": {
                    "description": "InfringementApproval",
                    "type": "object"
                },
                "infringementId": {
                    "type": "number"
                },
                "infringementLogs": {
                    "description": "InfringementLog[]",
                    "type": "object"
                },
                "infringementRevisionHistory": {
                    "description": "InfringementRevisionHistory[]",
                    "type": "object"
                },
                "issuer": {
                    "description": "Issuer",
                    "type": "object"
                },
                "issuerStatus": {
                    "type": "string"
                },
                "issuerStatusDescription": {
                    "type": "string"
                },
                "lastStatusChangeDate": {
                    "type": "string"
                },
                "lastSuccessfulPayment": {
                    "description": "Payment",
                    "type": "object"
                },
                "latestPaymentDate": {
                    "type": "string"
                },
                "location": {
                    "description": "PhysicalLocation",
                    "type": "object"
                },
                "logs": {
                    "description": "Log[]",
                    "type": "object"
                },
                "nomination": {
                    "description": "Nomination",
                    "type": "object"
                },
                "notes": {
                    "description": "InfringementNote[]",
                    "type": "object"
                },
                "noticeNumber": {
                    "type": "string"
                },
                "notifications": {
                    "description": "Notification[]",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "offenceDate": {
                    "type": "string"
                },
                "originalAmount": {
                    "type": "string"
                },
                "payments": {
                    "description": "Payment[]",
                    "type": "object"
                },
                "penaltyAmount": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reasonCode": {
                    "type": "string"
                },
                "status": {
                    "default": "Due",
                    "enum": [
                        "Missing Contract",
                        "Missing Vehicle",
                        "Valid"
                    ],
                    "type": "string"
                },
                "systemStatus": {
                    "default": "Valid",
                    "enum": [
                        "Missing Contract",
                        "Missing Vehicle",
                        "Valid"
                    ],
                    "type": "string"
                },
                "totalAmount": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "Parking",
                        "Traffic",
                        "Environmental",
                        "Other"
                    ],
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicle": {
                    "description": "Vehicle",
                    "type": "object"
                }
            },
            "required": [
                "createdAt",
                "updatedAt",
                "infringementId",
                "noticeNumber",
                "offenceDate",
                "latestPaymentDate",
                "reason",
                "reasonCode",
                "amountDue",
                "originalAmount",
                "totalAmount",
                "penaltyAmount",
                "status",
                "issuerStatus",
                "issuerStatusDescription",
                "systemStatus",
                "caseNumber",
                "brn",
                "externalChangeDate",
                "lastStatusChangeDate",
                "approvedDate",
                "type",
                "lastSuccessfulPayment",
                "vehicle",
                "contract",
                "issuer",
                "nomination",
                "payments",
                "logs",
                "notifications",
                "document",
                "location",
                "infringementLogs",
                "notes",
                "infringementRevisionHistory",
                "infringementApproval"
            ],
            "type": "object"
        },
...
        "OcrDetails": {
            "title": "OcrDetails",
            "properties": {
                "car": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            },
            "required": [
                "contract",
                "customer",
                "owner",
                "car",
                "start",
                "end"
            ],
            "type": "object"
        },
...
        "UpdateDocumentDto": {
            "title": "UpdateDocumentDto",
            "properties": {
                "fileName": {
                    "type": "string"
                }
            },
            "required": [
                "fileName"
            ],
            "type": "object"
        },
        "UpdateInfringementDto": {
            "title": "UpdateInfringementDto",
            "properties": {
                "amountDue": {
                    "type": "string"
                },
                "approvedDate": {
                    "type": "string"
                },
                "brn": {
                    "type": "string"
                },
                "caseNumber": {
                    "type": "string"
                },
                "dateRedirectionCompleted": {
                    "description": "The date that the redirection completed",
                    "type": "string"
                },
                "infringementStatus": {
                    "enum": [
                        "Due",
                        "Outstanding",
                        "Paid",
                        "Closed",
                        "Approved for Payment"
                    ],
                    "type": "string"
                },
                "isExternal": {
                    "type": "boolean"
                },
                "issuer": {
                    "type": "string"
                },
                "issuerStatus": {
                    "type": "string"
                },
                "issuerStatusDescription": {
                    "type": "string"
                },
                "latestPaymentDate": {
                    "type": "string"
                },
                "nominationStatus": {
                    "enum": [
                        "Pending",
                        "Acknowledged",
                        "In Redirection Process",
                        "Closed",
                        "Redirection Completed"
                    ],
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "notes": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "noticeNumber": {
                    "type": "string"
                },
                "offenceDate": {
                    "type": "string"
                },
                "originalAmount": {
                    "type": "string"
                },
                "paymentAmount": {
                    "type": "number"
                },
                "paymentDate": {
                    "type": "string"
                },
                "paymentReference": {
                    "type": "string"
                },
                "reason": {
                    "description": "The reason for the offence",
                    "type": "string"
                },
                "reasonCode": {
                    "description": "The reason code for the offence",
                    "type": "string"
                },
                "redirectionIdentifier": {
                    "description": "An identifier of the person or business that you are redirecting the infringement to",
                    "type": "string"
                },
                "redirectionLetterSendDate": {
                    "description": "The date that the redirection letter was sent to the municipality",
                    "type": "string"
                },
                "redirectionReference": {
                    "description": "A reference attached to the redirection",
                    "type": "string"
                },
                "redirectionType": {
                    "enum": [
                        "Manual Email",
                        "ATG Integration",
                        "Manual Upload",
                        "External",
                        "Telaviv",
                        "Jerusalem",
                        "Police",
                        "Mileon"
                    ],
                    "type": "string"
                },
                "setRedirectionCompletionDate": {
                    "type": "boolean"
                },
                "setRedirectionIdentifier": {
                    "type": "boolean"
                },
                "timezone": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "Parking",
                        "Traffic",
                        "Environmental",
                        "Other"
                    ],
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
...
}